Move Network, Magic, and magics to zebra-chain
