+++ zebra-chain/Cargo.toml
@@ -16,6 +16,7 @@ hex = "0.4"
 lazy_static = "1.4.0"
 ripemd160 = "0.8.0"
 secp256k1 = "0.17.2"
+serde = { version = "1", features = ["serde_derive"] }
 sha2 = "0.8"
 thiserror = "1"
 # ZF deps
