+++ zebra-chain/src/network.rs
@@ -1,4 +1,6 @@
-use crate::{constants::magics, protocol::external::types::Magic};
+//! Network-specific types.
+
+use crate::types::Magic;
 
 /// An enum describing the possible network choices.
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
@@ -17,4 +19,13 @@ impl Network {
             Network::Testnet => magics::TESTNET,
         }
     }
+}
+
+/// Magic numbers used to identify different Zcash networks.
+pub mod magics {
+    use super::*;
+    /// The production mainnet.
+    pub const MAINNET: Magic = Magic([0x24, 0xe9, 0x27, 0x64]);
+    /// The testnet.
+    pub const TESTNET: Magic = Magic([0xfa, 0x1a, 0xf9, 0xbf]);
 }
