+++ zebra-network/src/constants.rs
@@ -40,15 +40,6 @@ pub const CURRENT_VERSION: Version = Version(170_009);
 /// The minimum version supported for peer connections.
 pub const MIN_VERSION: Version = Version(170_009);
 
-/// Magic numbers used to identify different Zcash networks.
-pub mod magics {
-    use super::*;
-    /// The production mainnet.
-    pub const MAINNET: Magic = Magic([0x24, 0xe9, 0x27, 0x64]);
-    /// The testnet.
-    pub const TESTNET: Magic = Magic([0xfa, 0x1a, 0xf9, 0xbf]);
-}
-
 #[cfg(test)]
 mod tests {
